Feat: Login feito

+++ src/main/resources/db/migration/V3__create-table-fornecedores.sql
@@ -1,7 +1,7 @@
 CREATE TABLE fornecedores (
-    id SERIAL PRIMARY KEY,
+    id BIGSERIAL PRIMARY KEY,
     nome VARCHAR(255) NOT NULL,
     cnpj VARCHAR(18) UNIQUE NOT NULL,
-    usuario_id INT NOT NULL UNIQUE,
+    usuario_id BIGINT NOT NULL UNIQUE,
     CONSTRAINT fk_usuario_fornecedor FOREIGN KEY (usuario_id) REFERENCES usuarios (id) ON DELETE CASCADE
 );
